requirements.txt
pandas==2.2.0
python-dotenv==1.0.0
google-genai==0.3.0